Extract runtime control watcher into agent/runtime/control_watcher.rs
